+++ lib/honeybadger/rack/error_notifier.rb
@@ -21,9 +21,9 @@ module Honeybadger
     class ErrorNotifier
       extend Forwardable
 
-      def initialize(app, config)
+      def initialize(app, config = nil)
         @app = app
-        @config = config
+        @config = config || Agent.config
       end
 
       def call(env)