+++ lib/honeybadger/rack/metrics_reporter.rb
@@ -1,5 +1,6 @@
 require 'honeybadger'
 
+# Deprecated
 module Honeybadger
   module Rack
     class MetricsReporter