+++ lib/honeybadger/rack/user_feedback.rb
@@ -2,6 +2,8 @@ require 'erb'
 require 'uri'
 require 'forwardable'
 
+require 'honeybadger/agent'
+
 begin
   require 'i18n'
 rescue LoadError
@@ -19,9 +21,9 @@ module Honeybadger
     class UserFeedback
       extend Forwardable
 
-      def initialize(app, config)
+      def initialize(app, config = nil)
         @app = app
-        @config = config
+        @config = config || Agent.config
       end
 
       def call(env)